requirements.txt

# --------- pytorch --------- #
torch==1.12.0
# torchvision>=0.11.0
pytorch-lightning==1.7.3
torchmetrics>=0.9.3
torchtext
torchdata

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
tensorboard

# --------- linters --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- others --------- #
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
pudb            # debugger
lmdb
atom3d
scikit-optimize
numba

# --------- project related --------- #
biopython==1.79
einops
debugpy
matplotlib
pandas
seaborn
opt_einsum
sympy
e3nn
fairscale
# --------- jupyter --------- #
jupyter
ipykernel
